+++ app/(main)/favorites/page.tsx
@@ -1,10 +1,13 @@
 import FavoritesClient from "@/app/components/pages/main/favorites/FavoritesClient";
 import ClientOnly from "@/app/components/ClientOnly";
+import { ProtectedRoute } from "@/app/context/AuthContext";
 
 const FavoritesPage = async () => {
   return (
     <ClientOnly>
-      <FavoritesClient />
+      <ProtectedRoute>
+        <FavoritesClient />
+      </ProtectedRoute>
     </ClientOnly>
   );
 };