+++ app/(main)/reservations/review/[reservationId]/page.tsx
@@ -1,5 +1,6 @@
 import ReviewClient from "@/app/components/pages/main/reservations/review/ReviewClient";
 import ClientOnly from "@/app/components/ClientOnly";
+import { ProtectedRoute } from "@/app/context/AuthContext";
 
 interface IParams {
   reservationId: string;
@@ -8,7 +9,9 @@ interface IParams {
 const ReviewPage = async ({ params }: { params: IParams }) => {
   return (
     <ClientOnly>
-      <ReviewClient reservationId={params.reservationId} />
+      <ProtectedRoute>
+        <ReviewClient reservationId={params.reservationId} />
+      </ProtectedRoute>
     </ClientOnly>
   );
 };