+++ app/components/modals/FindModal.tsx
@@ -116,6 +116,8 @@ const FindModal = () => {
         <Calendar
           onChange={(value) => setDateRange(value.selection)}
           value={dateRange}
+          direction={"horizontal"}
+          displayTwoMonths
         />
       </div>
     );