+++ app/context/AuthContext.tsx
@@ -2,10 +2,11 @@
 
 import React, { createContext, useContext, useEffect, useState } from "react";
 import { isAuthenticated as isAuthenticatedFetch } from "pawpal-fe-common/context";
+import Loader from "../components/Loader";
+import { redirect } from "next/navigation";
 
 interface AuthContextType {
-  isAuthenticated: boolean;
-  authStatus: boolean;
+  authStatus: boolean | null;
   setAuthStatus: (status: boolean) => void;
 }
 
@@ -14,30 +15,26 @@ const AuthContext = createContext<AuthContextType | undefined>(undefined);
 export const AuthProvider: React.FC<{ children: React.ReactNode }> = ({
   children,
 }) => {
-  const [isAuthenticated, setIsAuthenticated] = useState<boolean>(false);
-
-  const authStatus = isAuthenticated;
+  const [authStatus, setAuthStatus] = useState<boolean | null>(null);
 
   useEffect(() => {
     const fetchInitialAuthStatus = async () => {
       const status = await isAuthenticatedFetch();
-      setIsAuthenticated(status);
+      setAuthStatus(status);
     };
 
     fetchInitialAuthStatus();
   }, []);
 
-  const setAuthStatus = (status: boolean) => {
-    setIsAuthenticated(status);
-  };
+  if (authStatus === false || authStatus === true) {
+    return (
+      <AuthContext.Provider value={{ authStatus, setAuthStatus }}>
+        {children}
+      </AuthContext.Provider>
+    );
+  }
 
-  return (
-    <AuthContext.Provider
-      value={{ isAuthenticated, authStatus, setAuthStatus }}
-    >
-      {children}
-    </AuthContext.Provider>
-  );
+  return <Loader />;
 };
 
 // Hook to use Auth Context
@@ -47,4 +44,20 @@ export const useAuth = () => {
     throw new Error("useAuth must be used within an AuthProvider");
   }
   return context;
+};
+
+export const ProtectedRoute = ({ children }: { children: React.ReactNode }) => {
+  const { authStatus } = useAuth();
+
+  useEffect(() => {
+    if (!authStatus) {
+      return redirect("/auth");
+    }
+  }, [authStatus]);
+
+  if (!authStatus) {
+    return null;
+  }
+
+  return children;
 };